+++ fundingmate-front/src/Component/Reward/Finish/Finish.js
@@ -1,5 +1,5 @@
 import React from "react";
-import "../Reward.css";
+import "../../../Page/Reward/Reward.css";
 
 const Finish = () => {
   return (
