--- fundingmate-front/src/Component/Reward/Reward.js
@@ -1,18 +1,0 @@
-import "../../App.css";
-import Banner from "./Banner/Banner";
-import Rewarding from "./Rewarding/Rewarding";
-import Preopen from "./Preopen/Preopen";
-import Finish from "./Finish/Finish";
-
-const Reward = () => {
-  return (
-    <div className="App">
-      <Banner />
-      <Rewarding />
-      <Preopen />
-      <Finish />
-    </div>
-  );
-};
-
-export default Reward;
