+++ fundingmate-front/src/Component/RewardDetail/Detail/Contact/Contact.js
@@ -1,4 +1,5 @@
 import React from "react";
+import "../../../../Page/Rewarddetail/Rewarddetail.css";
 
 const Contact = () => {
   return <div>Contact</div>;
