+++ fundingmate-front/src/Component/RewardDetail/Detail/Guide/Guide.js
@@ -1,4 +1,5 @@
 import React from "react";
+import "../../../../Page/Rewarddetail/Rewarddetail.css";
 
 const Guide = () => {
   return <div>Guide</div>;
