+++ fundingmate-front/src/Component/RewardDetail/Detail/Story/Story.js
@@ -1,4 +1,5 @@
 import React from "react";
+import "../../../../Page/Rewarddetail/Rewarddetail.css";
 
 const Story = () => {
   return <div>Story</div>;
