--- fundingmate-front/src/Component/RewardDetail/Rewarddetail.css
@@ -1,78 +1,0 @@
-.desc_subtitle {
-  font-size: 11px;
-  margin-top: 1%;
-  margin-bottom: 0.2%;
-  margin-left: 2.5%;
-}
-
-b {
-  color: var(--main-color);
-}
-
-.desc_title {
-  font-size: 15px;
-  margin-bottom: 0.8%;
-  margin-left: 2.4%;
-}
-
-.desc_contents {
-  display: flex;
-}
-
-.main_img {
-  width: 490px;
-  height: 300px;
-}
-
-.desc_img {
-  text-align: center;
-}
-.sub_img1 {
-  width: 100px;
-  height: 52.583px;
-  margin-right: 0.7%;
-}
-
-.sub_img2 {
-  width: 100px;
-  height: 52.583px;
-}
-
-.fund_category {
-  color: var(--main-color);
-  font-size: 13px;
-  margin-bottom: 0.5%;
-}
-
-.fund_price {
-  font-size: 20px;
-}
-
-.rewarding {
-  color: var(--black);
-  font-size: 12px;
-}
-
-.fund_rate {
-  display: flex;
-  width: 60%;
-  margin-left: 0.6%;
-  color: var(--grey);
-  font-size: 12px;
-  margin-top: 1%;
-}
-
-.fund_rate_title {
-  width: 68%;
-}
-
-.fund_rate_per {
-  width: 101%;
-  color: var(--black);
-}
-
-.fund_rate_price {
-  color: var(--black);
-  font-size: 10px;
-  margin-top: 1%;
-}
