--- fundingmate-front/src/Component/RewardDetail/Rewarddetail.js
@@ -1,18 +1,0 @@
-import React from "react";
-import Desc from "./Desc/Desc";
-import Contact from "./Detail/Contact/Contact";
-import Guide from "./Detail/Guide/Guide";
-import Story from "./Detail/Story/Story";
-
-const Rewarddetail = () => {
-  return (
-    <div className="App">
-      <Desc />
-      <Contact />
-      <Guide />
-      <Story />
-    </div>
-  );
-};
-
-export default Rewarddetail;
