+++ fundingmate-front/src/Page/Rewarddetail/Rewarddetail.css
@@ -1,0 +1,148 @@
+.desc_subtitle {
+  font-size: 11px;
+  margin-top: 1%;
+  margin-bottom: 0.2%;
+  margin-left: 2.5%;
+}
+
+b {
+  color: var(--main-color);
+}
+
+.desc_title {
+  font-size: 15px;
+  margin-bottom: 0.8%;
+  margin-left: 2.4%;
+}
+
+.desc_contents {
+  display: flex;
+}
+
+.main_img {
+  width: 490px;
+  height: 300px;
+}
+
+.desc_img {
+  text-align: center;
+}
+.sub_img1 {
+  width: 100px;
+  height: 52.583px;
+  margin-right: 0.7%;
+}
+
+.sub_img2 {
+  width: 100px;
+  height: 52.583px;
+}
+
+.fund_category {
+  color: var(--main-color);
+  font-size: 13px;
+  margin-bottom: 0.5%;
+}
+
+.fund_price {
+  font-size: 20px;
+}
+
+.rewarding {
+  color: var(--black);
+  font-size: 12px;
+}
+
+.fund_rate {
+  display: flex;
+  width: 60%;
+  margin-left: 0.6%;
+  color: var(--grey);
+  font-size: 12px;
+  margin-top: 2%;
+}
+
+.fund_rate_title {
+  width: 68%;
+}
+
+.fund_rate_per {
+  width: 101%;
+  color: var(--black);
+}
+
+.fund_rate_price {
+  color: var(--black);
+  font-size: 10px;
+  margin-top: 1%;
+}
+
+.fund_date {
+  display: flex;
+  width: 85%;
+  margin-left: 0.6%;
+  color: var(--grey);
+  font-size: 12px;
+  margin-top: 2%;
+}
+
+.fund_date_title {
+  width: 44%;
+}
+
+.fund_date_dday {
+  color: var(--black);
+  width: 43%;
+}
+
+.fund_date_enddate {
+  color: var(--black);
+  font-size: 10px;
+  margin-top: 1%;
+}
+
+.fund_people {
+  display: flex;
+  width: 37%;
+  margin-left: 0.6%;
+  color: var(--grey);
+  font-size: 12px;
+  margin-top: 2%;
+  justify-content: center;
+  justify-items: center;
+}
+
+.fund_people_title {
+  width: 44%;
+}
+
+.fund_people_count {
+  color: var(--black);
+  width: 43%;
+}
+
+.fund_btn {
+  width: 70%;
+  background-color: var(--main-color);
+  color: var(--white);
+  border: none;
+  border-radius: 5px;
+  font-size: 12px;
+  height: 20px;
+}
+
+.proj_share {
+  margin-top: 2%;
+  width: 40%;
+  background-color: var(--white);
+  border: 1px solid #000;
+  font-size: 12px;
+  height: 25px;
+  border-radius: 5px;
+  transition: all 0.2s linear;
+}
+
+.proj_share:hover {
+  border: 1px solid #14b8ff;
+  color: #14b8ff;
+}
