+++ fundingmate-front/src/Page/Rewarddetail/Rewarddetail.js
@@ -1,0 +1,17 @@
+import React from "react";
+import Desc from "../../Component/RewardDetail/Desc/Desc";
+import Contact from "../../Component/RewardDetail/Detail/Contact/Contact";
+import Guide from "../../Component/RewardDetail/Detail/Guide/Guide";
+import Story from "../../Component/RewardDetail/Detail/Story/Story";
+const Rewarddetail = () => {
+  return (
+    <div className="App">
+      <Desc />
+      <Contact />
+      <Guide />
+      <Story />
+    </div>
+  );
+};
+
+export default Rewarddetail;
