+++ fundingmate-front/src/index.css
@@ -11,6 +11,7 @@
   --black: #000;
   --sub-color: #ffce50;
   --grey: #9b9b9b;
+  --white: white;
 }
 
 * {
